Just to see what not understanding inheritance means.
Also, a ticket inherits the name and the date of the event.

But it does not inherit the identifier of a seat...

you create the seat on the main stack then you give a copy of it to the row class where you decide how many copies of the seat you give. but then, you cannot initialise the identifier by setter or by constructor for each seat in the row class. which is kinda weird. But would you inherit a vector of locations, then a vector of zones (an array of pointers to pointers), then of rows then of seats for ONLY one ticket???? idts